+++ src/main/java/de/maxhenkel/easyvillagers/blocks/tileentity/render/AutoTraderRenderer.java
@@ -1,21 +1,13 @@
 package de.maxhenkel.easyvillagers.blocks.tileentity.render;
 
 import com.mojang.blaze3d.matrix.MatrixStack;
-import de.maxhenkel.corelib.client.RenderUtils;
-import de.maxhenkel.easyvillagers.blocks.AutoTraderBlock;
 import de.maxhenkel.easyvillagers.blocks.tileentity.AutoTraderTileentity;
-import net.minecraft.block.BlockState;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.BlockRendererDispatcher;
 import net.minecraft.client.renderer.IRenderTypeBuffer;
-import net.minecraft.client.renderer.RenderTypeLookup;
 import net.minecraft.client.renderer.entity.VillagerRenderer;
 import net.minecraft.client.renderer.tileentity.TileEntityRenderer;
 import net.minecraft.client.renderer.tileentity.TileEntityRendererDispatcher;
 import net.minecraft.resources.IReloadableResourceManager;
-import net.minecraft.util.Direction;
-import net.minecraft.util.math.vector.Vector3f;
-import net.minecraftforge.client.model.data.EmptyModelData;
 
 public class AutoTraderRenderer extends TileEntityRenderer<AutoTraderTileentity> {
 
@@ -30,46 +22,10 @@ public class AutoTraderRenderer extends TileEntityRenderer<AutoTraderTileentity>
 
     @Override
     public void render(AutoTraderTileentity trader, float partialTicks, MatrixStack matrixStack, IRenderTypeBuffer buffer, int combinedLightIn, int combinedOverlayIn) {
-        matrixStack.push();
-
         if (renderer == null) {
             renderer = new VillagerRenderer(minecraft.getRenderManager(), (IReloadableResourceManager) minecraft.getResourceManager());
         }
-
-        Direction direction = Direction.SOUTH;
-        if (!trader.isFakeWorld()) {
-            direction = trader.getBlockState().get(AutoTraderBlock.FACING);
-        }
-
-        if (trader.getVillagerEntity() != null) {
-            matrixStack.push();
-
-            matrixStack.translate(0.5D, 1D / 16D, 0.5D);
-            matrixStack.rotate(Vector3f.YP.rotationDegrees(-direction.getHorizontalAngle()));
-            matrixStack.translate(0D, 0D, -4D / 16D);
-            matrixStack.scale(0.45F, 0.45F, 0.45F);
-            renderer.render(trader.getVillagerEntity(), 0F, 1F, matrixStack, buffer, combinedLightIn);
-            matrixStack.pop();
-        }
-
-        if (trader.hasWorkstation()) {
-            matrixStack.push();
-
-            matrixStack.translate(0.5D, 1D / 16D, 0.5D);
-            matrixStack.rotate(Vector3f.YP.rotationDegrees(-direction.getHorizontalAngle()));
-            matrixStack.translate(0D, 0D, 2D / 16D);
-            matrixStack.translate(-0.5D, 0D, -0.5D);
-            matrixStack.scale(0.45F, 0.45F, 0.45F);
-            matrixStack.translate(0.5D / 0.45D - 0.5D, 0D, 0.5D / 0.45D - 0.5D);
-
-            BlockState workstation = trader.getWorkstation().getDefaultState();
-            BlockRendererDispatcher dispatcher = minecraft.getBlockRendererDispatcher();
-            int color = minecraft.getBlockColors().getColor(workstation, null, null, 0);
-            dispatcher.getBlockModelRenderer().renderModel(matrixStack.getLast(), buffer.getBuffer(RenderTypeLookup.func_239221_b_(workstation)), workstation, dispatcher.getModelForState(workstation), RenderUtils.getRed(color), RenderUtils.getGreen(color), RenderUtils.getBlue(color), combinedLightIn, combinedOverlayIn, EmptyModelData.INSTANCE);
-            matrixStack.pop();
-        }
-
-        matrixStack.pop();
+        TraderRenderer.renderTraderBase(minecraft, renderer, trader, partialTicks, matrixStack, buffer, combinedLightIn, combinedOverlayIn);
     }
 
 }